requirements.txt
absl-py==1.2.0
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
black==22.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
easyocr==1.6.0
flatbuffers==1.12
fonttools==4.37.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h11==0.13.0
h5py==3.7.0
idna==3.3
imageio==2.21.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
joblib==1.1.0
jsonmerge==1.8.0
jsonschema==4.14.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mypy-extensions==0.4.3
networkx==2.8.6
ninja==1.10.2.3
nltk==3.7
numpy==1.23.2
oauthlib==3.2.0
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
Shapely==1.8.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
textblob==0.17.1
threadpoolctl==3.1.0
tifffile==2022.8.12
tomli==2.0.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.3.0
urllib3==1.26.12
webcolors==1.12
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.8.1
